CriticalTransitions.jl

CriticalTransitions — Module

CriticalTransitions.jl

CriticalTransitions.jl is a first-of-its-kind software for formalizing, automating, and making extendable, the analysis of critical transitions in dynamical systems. Current content exists along two independent paths: noise- and rate- induced transitions.

The main software highlights are:

easily construct stochastic and nonautonomous dynamical systems

efficiently sample transition path ensembles

calculate minimum action paths and critical forcing rates

use a growing toolbox of tested and documented functions implementing concepts of large deviation theory, transition path theory, and rate-induced tipping

and more features shown in the documentation and planned for the future!

CriticalTransitions.jl can be used as a standalone package, or as part of DynamicalSystems.jl.

To install it, run import Pkg; Pkg.add("CriticalTransitions").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Developers: Reyk Börner, Orjan Ameye, Ryan Deeley, Raphael Römer and George Datseris

Thanks to Jeroen Wouters, Calvin Nesbitt, Tobias Grafke and Oliver Mehling

This package was created as part of the CriticalEarth project.

Latest news

Upcoming release!

Getting started

See the Tutorial, and after that either consult the API page for all the available functionality or visit individual examples!

People

Main developers:

Reyk Börner (@reykboerner)

Orjan Ameye (@oameye)

Ryan Deeley (@ryandeeley)

Raphael Römer (@raphael-roemer)

George Datseris (@Datseris)

Thanks to Jeroen Wouters, Calvin Nesbitt, Tobias Grafke & Oliver Mehling.

This package got started in the EU-funded CriticalEarth project.