Analysing sample transition paths

CriticalTransitions.fw_action — Method

fw_action(sys::StochSystem, path, time)

Calculates the Freidlin-Wentzell action of a given path with time points time in a drift field specified by the deterministic dynamics of sys.

The path must be a (D x N) matrix, where D is the dimensionality of the system sys and N is the number of path points. The time array must have length N.

Returns a single number, which is the value of the action integral

$S_T[\phi_t] = \frac{1}{2} \int_0^T || \dot \phi_t - b ||^2 dt$

where $\phi_t$ denotes the path and $b$ the drift field.

Missing docstring.

Missing docstring for om_action(sys::StochSystem, path, time). Check Documenter's build log for details.