Simulating the system

We provide two main functions to simulate a CoupledSDEs forward in time:

relax for the system's deterministic evolution (in the absence of noise) based on the ODEProblem of DifferentialEquations.jl

simulate for stochastic dynamics based on the SDEProblem of DifferentialEquations.jl

Deterministic dynamics

CriticalTransitions.relax — Method

relax(sys::CoupledSDEs, T; ...) -> Any
relax(sys::CoupledSDEs, T, init; alg, kwargs...) -> Any

Simulates the deterministic dynamics of CoupledSDEs sys in time with an amount T, starting at initial condition init.

This function integrates sys.f forward in time, using the ODEProblem functionality of DifferentialEquations.jl. Thus, relax is identical to simulate when setting the noise strength sys.σ = 0.

Keyword arguments

solver=Tsit5(): ODE solver. Defaults to Tsitouras 5/4 Runge-Kutta method.

kwargs...: keyword arguments for solve(ODEProblem)

For more info, see ODEProblem. For stochastic integration, see simulate.

Stochastic dynamics

CriticalTransitions.simulate — Method

simulate(sys::CoupledSDEs, T; ...) -> Any
simulate(sys::CoupledSDEs, T, init; alg, kwargs...) -> Any

Simulates the CoupledSDEs sys forward in time with an amount T, starting at initial condition init.

This function uses the SDEProblem functionality of DifferentialEquations.jl.

Keyword arguments

alg=SOSRA(): SDE solver. Defaults to an adaptive strong order 1.5 method.

kwargs...: keyword arguments for solve(SDEProblem)

For more info, see SDEProblem.