Analyzing a system's stability properties

To use the following functionalities, you need to load ChoasTools.jl and Attractors.jl.

Fixed points

CriticalTransitions.equilib — Method

equilib(sys::CoupledSDEs, state; kwargs...)

Returns the equilibrium solution of the system sys for given initial condition state.

Warning: This algorithm simply evolves the deterministic system forward in time until a steady-state condition is satisfied. Thus, the algorithm may output a false solution if it gets stuck in a quasi-equilibrium, or slowly evolving state. For more robust results, use fixedpoints.

Keyword arguments:

abstol = 1e-5: steady-state condition. Simulation ends when the rate of change (Euclidean distance in state space) of the state falls below abstol.

tmax = 1e5: maximum simulation time before the algorithm stops even if the steady-state condition is not reached.

dt = 0.01: time step of the ODE solver.

solver = Euler(): ODE solver used for evolving the state.

ChaosTools.fixedpoints — Function

fixedpoints(
    sys::CoupledSDEs,
    bmin::Vector,
    bmax::Vector
) -> Tuple{StateSpaceSet, Vector{Vector{ComplexF64}}, Vector{Bool}}

Returns fixed points, their eigenvalues and stability of the system sys within the state space volume defined by bmin and bmax.

This is a wrapper around the fixedpoints function of DynamicalSystems.jl.

Input

bmin (Vector): lower limits of the state space box to be considered, as a vector of coordinates

bmax (Vector): upper limits

alternatively box (IntervalBox) can replace bmin and bmax

Example: fixedpoints(sys, [-2,-1,0], [2,1,1]) finds the fixed points of the 3D system sys in a cube defined by the intervals [-2,2] × [-1,1] × [0,1].

Output

[fp, eigs, stable]

fp: Dataset of fixed points

eigs: vector of Jacobian eigenvalues of each fixed point

stable: vector of booleans indicating the stability of each fixed point (true=stable, false=unstable)

Additional methods

fixedpoints(sys::CoupedSDEs, box)

Basins of attraction

CriticalTransitions.basins — Function

basins(
    sys::CoupledSDEs,
    A,
    B,
    C,
    H;
    bstep,
    pstep,
    ϵ_mapper,
    showprogress,
    kwargs...
) -> Any

Computes the basins of attraction of CoupledSDEs sys on a plane spanned by the distinct points A, B, C and limited by the box H. Uses the AttractorsViaProximity function from DynamicalSystems.jl to compute the basins of attraction.

A, B, C are elements of $\mathbb{R}^d$ (where $d$ is the dimension of the sys) and H is a hyperrectangle in $\mathbb{R}^d$.

The plane is given by $P_{U,V} := \{A+u(B-A)+v(C-A)\in\mathbb{R}^d: u \in U,\, v\in V\}$ for some closed and bounded real intervals $U$ and $V$ which are selected such that both

$P_{U,\,V} \subseteq H$, and

$U\times V\subseteq\mathbb{R}^2$ has maximal area,

i.e. $P_{U,\,V}$ is the "largest" possible plane contained within H. This plane is determined behind the scenes.

This function returns a four-dimensional vector. The first two entries are discretised versions of the interval $U$ and $V$ (as defined above, of lengths $\ell_U,\,\ell_V$ respectively); the third entry is a dictionary of the attractors (stable equilibria) of the system within H, and the final entry is an $\ell_V\times\ell_U$ matrix of integers that group the initial conditions (written in terms of $A+u(B-A)+v(C-A)$ where $u\in U$ and $v\in V$) by which attractor they will in time converge to.

Keyword arguments

bstep = [0.01, 0.01]: a vector of length two whose elements respectively specify the length of the incremental steps taken across each dimension in the discretisation of your plane

pstep = [0.1, 0.1]: a vector of length two whose elements give the increments of the mesh that the maximisation process of finding a plane from a box is taken over (for more information see the source code of the function plane in the src/systemanalysis/planeofbox.jl file)

ϵ_mapper = 0.01: ϵ parameter of AttractorsViaProximity

kwargs...: keyword arguments passed to the AttractorsViaProximity function (namely, Ttr, Δt, horizon_limit, mx_chk_lost)

CriticalTransitions.basinboundary — Function

basinboundary(X, Y, h; coords, A, B, C) -> Any

Computes the basin boundary for given output X, Y, h of the basins function.

To be further documented.

basinboundary(boa) -> Any

Computes the basin boundary for given output boa of the basins function.

To be further documented.

CriticalTransitions.basboundary — Function

basboundary(
    sys::CoupledSDEs,
    xrange::Vector,
    yrange::Vector,
    xspacing::Float64,
    attractors::Vector;
    eps1,
    ϵ_mapper,
    dt_mapper,
    solver,
    maxit
) -> Tuple{Any, Any}

This function computes the basin boundary.

Edge tracking

The edge tracking algorithm is a simple numerical method to find the edge state or (possibly chaotic) saddle on the boundary between two basins of attraction. It is first introduced by Battelino et al. (1988) and further described by Skufca et al. (2006).

Attractors.edgetracking — Function

edgetracking(
    sys::CoupledSDEs,
    u1,
    u2,
    attractors::Vector;
    eps1,
    eps2,
    converge,
    dt,
    tmax,
    ϵ_mapper,
    dt_mapper,
    solver,
    maxit,
    verbose,
    output_all,
    kwargs...
)

Runs the edge tracking algorithm.

Input

sys: dynamical system of type CoupledSDEs

u1, u2: initial states; must belong to different basins of attraction

attractors: vector of state vectors corresponding to the stable fixed points of sys

Keyword arguments

eps1 = 1e-9: tolerance for bisection distance

eps2 = 1e-8: tolerance for divergence of trajectories before re-bisecting

converge = 1e-5: convergence criterion for M state accuracy (Euclidean distance)

dt = 0.01: integration time step

tmax = Inf: maximum integration time of parallel trajectories until re-bisection

ϵ_mapper = 0.1: distance threshold for AttractorMapper

dt_mapper = 0.01: time step for AttractorMapper (keyword argument Δt)

solver = Vern9(): ODE solver from DifferentialEquations.jl

maxit = 100: maximum number of iterations before algorithm stops

verbose = true: print status updates during run

output_all = false: if false, returns M state, else returns all points of the track

kwargs...: additional keyword arguments of AttractorsViaProximity may be passed

Returns

If output_all, a single state vector corresponding to the found edge state is returned. Else, a triple edge, track1, track2 is returned, where track1 and track2 are the tracks along the edge within the basin of attraction of u1 and u2, respectively; edge is the track along the edge derived from (track1 + track2)/2.

Warning

May behave erroneously when run with solver = SimpleATsit5(), which is the default solver for AttractorsViaProximity. The recommended solver is Vern9().

Attractors.bisect_to_edge — Function

bisect_to_edge(
    sys::CoupledSDEs,
    u1,
    u2,
    attractors::Vector;
    eps1,
    ϵ_mapper,
    dt_mapper,
    solver,
    maxit,
    absto,
    relto,
    kwargs...
)

Bisects to the basin boundary between two initial points u1 and u2. Returns the two final points, one on each side of the basin boundary, that are less than eps1 apart from each other.

Keyword arguments

eps1=1e-9: tolerance for final distance between the two states

ϵ_mapper=0.1: ϵ parameter for AttractorMapper

dt_mapper = 0.01: time step for AttractorMapper (keyword argument Δt)

solver=Vern9(): solver for AttractorMapper

kwargs...: additional kwargs that can be passed to AttractorMapper