Convenience functions and types

Interface to DynamicalSystems.jl

Using the functionality of DynamicalSystems.jl

A CoupledSDEs can easily be turned into a CoupledODEs instance of DynamicalSystems.jl using the function CoupledODEs. Vice vera, a CoupledODEs system can be converted into a CoupledSDEs via CoupledSDEs(ds::CoupledODEs, g) with g the noise function.

This way, many of the methods in DynamicalSystems.jl can be used directly, even if we have not written an analogue method that takes CoupledSDEs as input. For example, the Lyapunov spectrum of a CoupledSDEs, here exemplified by the FitzHugh-Nagumo model, can be computed by typing:

using CriticalTransitions
using DynamicalSystems: lyapunovspectrum

function fitzhugh_nagumo(u, p, t)
    x, y = u
    ϵ, β, α, γ, κ, I = p

    dx = (-α * x^3 + γ * x - κ * y + I) / ϵ
    dy = -β * y + x

    return SA[dx, dy]
end

sys = CoupledSDEs(fitzhugh_nagumo, id_func, zeros(2), σ, [1.,3.,1.,1.,1.,0.])
ls = lyapunovspectrum(CoupledODEs(sys), 10000)

DynamicalSystemsBase.CoupledODEs — Method

CoupledODEs(sys::CoupledSDEs; diffeq, t0) -> CoupledODEs

Converts a CoupledSDEs into CoupledODEs from DynamicalSystems.jl.

CoupledSDEs utility functions

CriticalTransitions.noise_strength — Function

noise_strength(sys::CoupledSDEs) -> Any

Gives the noise strength specified in sys.

CriticalTransitions.covariance_matrix — Function

covariance_matrix(sys::CoupledSDEs) -> Any

Gives the covariance matrix specified in sys.

CriticalTransitions.noise_process — Function

noise_process(sys::CoupledSDEs) -> Any

Translates the stochastic process specified in sys into the language required by the SDEProblem of DynamicalSystems.jl.

CriticalTransitions.drift — Function

drift(sys::CoupledSDEs{IIP}, x) -> Any

Returns the drift field $b(x)$ of the CoupledSDEs sys at the state vector x.

CriticalTransitions.div_drift — Function

div_drift(sys::CoupledSDEs, x) -> Any

Computes the divergence of the drift field sys.f at the given point x.

Saving data

CriticalTransitions.make_jld2 — Function

make_jld2(
    text::String
) -> Union{JLD2.JLDFile{IOStream}, JLD2.JLDFile{JLD2.MmapIO}}
make_jld2(
    text::String,
    relpath::String
) -> Union{JLD2.JLDFile{IOStream}, JLD2.JLDFile{JLD2.MmapIO}}

Creates/opens a .jld2 file with filename of the format "ddmmyy_text.jld2". Relative file path specified by relpath (must end with /).

See make_h5 for generating a .h5 file.

CriticalTransitions.make_h5 — Function

make_h5(text::String) -> HDF5.File
make_h5(text::String, relpath::String) -> HDF5.File

Creates/opens a .h5 file with filename of the format "ddmmyy_text.h5". Relative file path specified by relpath (must end with /).

See make_jld2 for generating a .jld2 file.

length(sys.u)-dimensional box

CriticalTransitions.intervals_to_box — Method

intervals_to_box(bmin::Vector, bmax::Vector) -> Any

Generates a box from specifying the interval limits in each dimension.

bmin (Vector): lower limit of the box in each dimension

bmax (Vector): upper limit

Example

intervals_to_box([-2,-1,0], [2,1,1]) returns a 3D box of dimensions [-2,2] × [-1,1] × [0,1].