Functionality under development

Pathspace Langevin MCMC sampling for the FitzHugh-Nagumo system

CriticalTransitions.fhn_pathspace_sampling — Function

fhn_pathspace_sampling(ϵ, σ, T, dz; kwargs...)

Integrates the FitzHughNagumoSPDE in virtual time to generate pathspace Langevin MCMC samples.

Unless specified otherwise, the initial condition is a straight line between the fixed points R and L (for the default FHN parameters).

Arguments

ϵ (float): time scale parameter

σ (float): noise strength

T (float): total physical time

dz (float): physical time step

init=nothing: initial path. If nothing, init is the straight line between the attractors.

Keyword arguments

dt (float): virtual time step

tmax (float): total virtual time

a (float): first diagonal element of covariance matrix

start (string): Starting point of the path. Either "R" or "L" (default "R")

save_every (int): save the path at every save_every-th instance in virtual time

β, α, γ, κ, Ι: FHN parameters

Returns

A list of two matrices, each with columns = coordinates and rows = virtual time instances. The first matrix contains the u coordinates, the second matrix the v coordinates.

CriticalTransitions.FitzHughNagumoSPDE — Function

FitzHughNagumoSPDE(u, p, t)

System definition for the pathspace SPDE corresponding to the FitzHughNagumo system.

The parameter vector must contain the following (in this order):

ϵ (float): time scale parameter

β (float): FHN parameter

α (float): FHN parameter

γ (float): FHN parameter

κ (float): FHN parameter

Ι (float): FHN parameter

a (float): First diagonal element of covariance matrix

σ (float): noise strength

dz (float): physical time step