API

You guys need to fill this.

CriticalTransitions.transition — Function

transition(
    sys::CoupledSDEs,
    x_i,
    x_f;
    radii,
    tmax,
    radius_directions,
    cut_start,
    kwargs...
) -> Tuple{Any, Any, Bool}

Generates a sample transition from point x_i to point x_f.

This function simulates sys in time, starting from initial condition x_i, until entering a ball of given radius around x_f. If a seed was given to sys the solver is initialized with this seed. To change the seed you can pass a new seed to the seed keyword.

Keyword arguments

radii=(0.1, 0.1): radius of the ball around x_i and x_f, respectively

tmax=1e3: maximum time until the simulation stops even x_f has not been reached

radius_directions=1:length(sys.u): the directions in phase space to consider when calculating the radii rad_i and rad_f. Defaults to all directions. To consider only a subspace of state space, insert a vector of indices of the dimensions to be included.

cut_start=true: if false, returns the whole trajectory up to the transition

kwargs...: keyword arguments passed to CommonSolve.solve

Output

[path, times, success]

path (Matrix): transition path (size [dim × N], where N is the number of time points)

times (Vector): time values (since start of simulation) of the path points (size N)

success (bool): if true, a transition occurred (i.e. the ball around x_f has been reached), else false

See also transitions, trajectory.

Missing docstring.

Missing docstring for geometric_min_action_method. Check Documenter's build log for details.