Large deviation theory

Action functionals

Freidlin-Wentzell action

CriticalTransitions.fw_action — Method

fw_action(sys::StochSystem, path, time; kwargs...)

Calculates the Freidlin-Wentzell action of a given path with time points time in a drift field specified by the deterministic dynamics of sys.

The path must be a (D x N) matrix, where D is the dimensionality of the system sys and N is the number of path points. The time array must have length N.

Returns a single number, which is the value of the action functional

$S_T[\phi_t] = \frac{1}{2} \int_0^T || \dot \phi_t - b(\phi_t) ||^2_Q dt$

where $\phi_t$ denotes the path in state space, $b$ the drift field, and $T$ the total time of the path. The subscript $Q$ refers to the generalized norm $||a||_Q^2 := \langle a, Q^{-1} b \rangle$ (see anorm). Here $Q$ is the noise covariance matrix sys.Σ.

Keyword arguments

cov_inv = nothing: Inverse of the covariance matrix $\Sigma$. If nothing, $\Sigma^{-1}$ is computed automatically.

Geometric Freidlin-Wentzell action

CriticalTransitions.geometric_action — Function

geometric_action(sys::StochSystem, path, arclength=1; kwargs...)

Calculates the geometric action of a given path with specified arclength for the drift field sys.f.

For a given path $\varphi$, the geometric action $\bar S$ corresponds to the minimum of the Freidlin-Wentzell action $S_T(\phi)$ over all travel times $T>0$, where $\phi$ denotes the path's parameterization in physical time (see fw_action). It is given by the integral

$\bar S[\varphi] = \int_0^L \left( ||\varphi'||_Q \, ||b(\varphi)||_Q - \langle \varphi', \, b(\varphi) \rangle_Q \right) \, ds$

where $s$ is the arclength coordinate, $L$ the arclength, $b$ the drift field, and the subscript $Q$ refers to the generalized dot product $\langle a, b \rangle_Q := a^{\top} \cdot Q^{-1} b$ (see anorm). Here $Q$ is the noise covariance matrix sys.Σ.

Keyword arguments

cov_inv = nothing: Inverse of the covariance matrix $\Sigma$. If nothing, $\Sigma^{-1}$ is computed automatically.

Returns the value of the geometric action $\bar S$.

Onsager-Machlup action

CriticalTransitions.om_action — Method

om_action(sys::StochSystem, path, time; kwargs...)

Calculates the Onsager-Machlup action of a given path with time points time for the drift field sys.f at noise strength sys.σ.

The path must be a (D x N) matrix, where D is the dimensionality of the system sys and N is the number of path points. The time array must have length N.

Returns a single number, which is the value of the action functional

$I^{\sigma}_T[\phi_t] = \frac{1}{2} \int_0^T \left( || \dot \phi_t - b(\phi_t) ||^2_Q + \frac{\sigma^2}{2} \div(b) \right) \, dt$

where $\phi_t$ denotes the path in state space, $b$ the drift field, $T$ the total time of the path, and $\sigma$ the noise strength. The subscript $Q$ refers to the generalized norm $||a||_Q^2 := \langle a, Q^{-1} b \rangle$ (see anorm). Here $Q$ is the noise covariance matrix sys.Σ.

Keyword arguments

cov_inv = nothing: Inverse of the covariance matrix $\Sigma$. If nothing, $\Sigma^{-1}$ is computed automatically.

For convenience, a general action function is available where the type of functional is set as an argument:

CriticalTransitions.action — Method

action(sys::StochSystem, path::Matrix, time, functional; kwargs...)

Computes the action functional specified by functional for a given StochSystem sys and path parameterized by time.

functional = "FW": Returns the Freidlin-Wentzell action (fw_action)

functional = "OM": Returns the Onsager-Machlup action (om_action)

Minimum action paths

We provide the following two methods to calculate instantons, or minimum action paths, between two states of a StochSystem.

Minimum action method (MAM)

Minimization of the Freidlin-Wentzell action using the L-BFGS algorithm of Optim.jl.

CriticalTransitions.min_action_method — Method

min_action_method(sys::StochSystem, x_i::State, x_f::State, N::Int, T::Real; kwargs...)

Runs the Minimum Action Method (MAM) to find the minimum action path (instanton) between an initial state x_i and final state x_f.

This algorithm uses the minimizers of the Optim package to minimize the Freidlin-Wentzell action functional (see fw_action) for the given StochSystem sys. The path is initialized as a straight line between x_i and x_f, parameterized in time via N equidistant points and total time T. Thus, the time step between discretized path points is $\Delta t = T/N$. To set an initial path different from a straight line, see the multiple dispatch method

min_action_method(sys::StochSystem, init::Matrix, T::Real; kwargs...).

The minimization can be performed in blocks to save intermediate results.

Keyword arguments

functional = "FW": type of action functional to minimize. Defaults to fw_action, alternative: om_action.

maxiter = 100: maximum number of iterations before the algorithm stops.

blocks = 1: number of iterative optimization blocks

method = LBFGS(): minimization algorithm (see Optim)

save_info = true: whether to save Optim information

verbose = true: whether to print Optim information during the run

showprogress = false: whether to print a progress bar

kwargs...: any keyword arguments from Optim.Options (see docs)

Output

If save_info, returns Optim.OptimizationResults. Else, returns only the optimizer (path). If blocks > 1, a vector of results/optimizers is returned.

Geometric minimum action method (gMAM)

Minimization of the geometric action following Heymann and Vanden-Eijnden, PRL (2008).

CriticalTransitions.geometric_min_action_method — Function

geometric_min_action_method(sys::StochSystem, x_i::State, x_f::State, arclength=1; kwargs...)

Computes the minimizer of the Freidlin-Wentzell action using the geometric minimum action method (gMAM). Beta version, to be further documented.

To set an initial path different from a straight line, see the multiple dispatch method

geometric_min_action_method(sys::StochSystem, init::Matrix, arclength::Float64; kwargs...).

Keyword arguments

N = 100: number of discretized path points

maxiter = 100: maximum number of iterations before the algorithm stops

converge = 1e-5: convergence threshold for absolute change in action

method = LBFGS(): choice of optimization algorithm (see below)

tau = 0.1: step size (used only if method = "HeymannVandenEijnden")

Optimization algorithms

The method keyword argument takes solver methods of the Optim.jl package; alternatively, the option solver = "HeymannVandenEijnden" uses the original gMAM algorithm[1].

1Heymann and Vanden-Eijnden, PRL (2008)