Large deviation theory

Action functionals

CriticalTransitions.fw_action — Method

fw_action(sys::StochSystem, path, time)

Calculates the Freidlin-Wentzell action of a given path with time points time in a drift field specified by the deterministic dynamics of sys.

The path must be a (D x N) matrix, where D is the dimensionality of the system sys and N is the number of path points. The time array must have length N.

Returns a single number, which is the value of the action integral

$S_T[\phi_t] = \frac{1}{2} \int_0^T || \dot \phi_t - b ||^2 dt$

where $\phi_t$ denotes the path and $b$ the drift field.

Missing docstring.

Missing docstring for om_action(sys::StochSystem, path, time). Check Documenter's build log for details.

Calculating instantons (minimum action paths)

CriticalTransitions.mam — Method

mam(sys::StochSystem, x_i::State, x_f::State, N::Int, T::Float64; kwargs...)

Runs the Minimum Action Method (MAM) to find the minimum action path (instanton) between an initial state x_i and final state x_f.

This algorithm uses the minimizers of the Optim package to minimize the Freidlin-Wentzell action functional (see fw_action) for the given StochSystem sys. The path is initialized as a straight line between x_i and x_f, parameterized in time via N equidistant points and total time T. Thus, the time step between discretized path points is Δt = T/N. The minimization is performed in consecutive blocks of a fixed number of iterations each.

Keyword arguments

blocks = 10: number of blocks

block_iterations = 10: number of iterations per block

method = LBFGS(): minimization algorithm (see Optim)

output = "MAP": if "MAP", outputs the final path; if "all", outputs everything

showprogress = true: whether to print a progress bar

Output

The output can be controlled via the output keyword argument.

Alternative input

mam(sys::StochSystem, init::Matrix, T::Float64; kwargs...)