Large deviation theory

This section applies results of large deviation theory (LDT), particularly action minimization problems for computing most probable transition paths in stochastic dynamical systems driven by weak noise. For a description of the theory, see Freidlin and Wentzell [8] and [9]. An overview of numerical methods applying LDT is given in Grafke and Vanden-Eijnden [10].

Info

The methods in this section apply to $D$-dimensional stochastic dynamical systems of the form

\[\text{d} \mathbf{x} = \mathbf{b} (\mathbf{x}) \text{d}t + \sigma \mathbf{\Sigma} \text{d}\mathbf{W}_t \,,\]

where the drift field $\mathbf{b}$ may be non-gradient but the noise term must consist of Gaussian noise ($\mathbf{W}_t$ is a $D$-dimensional vector of independent standard Wiener processes) and a constant covariance matrix $\mathbf{Q} = \mathbf{\Sigma}\mathbf{\Sigma}^\top$.

This is a special case of the broader class of noise types supported by CoupledSDEs.

Action minimizers

Several methods have been proposed to calculate transition paths that minimize a given action functional. In the weak-noise limit, this minimum action path (or instanton) corresponds to the most probable transition path. While the minimum action method (MAM) is the most basic version, it is often beneficial to minimize the geometric action via a time-independent version called gMAM. The problem can also be cast in a Hamiltonian form, implemented as simple gMAM (sgMAM), which can have numerical advantages.

These methods apply to non-gradient systems driven by Gaussian noise. In gradient systems, minimum action paths between attractors coincide with heteroclinic orbits, which can be computed via the so-called string method.

To summarize, the following methods are available:

Minimum action method (MAM)

Geometric minimum action method (gMAM)

Simple geometric minimum action method (sgMAM)

String method

Minimum action method (MAM)

Minimization of the specified action functional using the optimization algorithm of Optimization.jl. See also E et al. [11].

CriticalTransitions.min_action_method — Function

min_action_method(sys::ContinuousTimeDynamicalSystem, x_i, x_f, T::Real; kwargs...)

Minimizes an action functional to obtain a minimum action path (instanton) between an initial state x_i and final state x_f in phase space.

This algorithm uses the Optimization.jl package to minimize the specified action functional (either fw_action or om_action) for the system sys over paths connecting x_i to x_f in time T.

The path is initialized as a straight line between x_i and x_f, parameterized in time via N equidistant points and total time T. Thus, the time step between discretized path points is $\Delta t = T/N$. To set an initial path different from a straight line, see the multiple dispatch method

min_action_method(sys::ContinuousTimeDynamicalSystem, init::Matrix, T::Real; kwargs...).

Returns a MinimumActionPath object containing the optimized path and the action value.

Keyword arguments

functional = "FW": type of action functional to minimize. Defaults to fw_action, alternative: "OM" for om_action.

N = 100: number of path points to use for the discretization of the path.

noise_strength = nothing: noise strength for the action functional. Specify only if functional = "OM".

method = Optimisers.Adam(): minimization algorithm (see Optimization.jl)

ad_type = Optimization.AutoFiniteDiff(): type of automatic differentiation to use (see Optimization.jl)

maxiters = 100: maximum number of iterations before the algorithm stops.

abstol=1e-8: absolute tolerance of action gradient to determine convergence

reltol=1e-8: relative tolerance of action gradient to determine convergence

verbose = true: whether to print Optimization information during the run

show_progress = false: whether to print a progress bar

min_action_method(sys::ContinuousTimeDynamicalSystem, init::Matrix, T::Real; kwargs...)

Minimizes the specified action functional to obtain a minimum action path (instanton) between fixed end points given a system sys and total path time T.

The initial path init must be a matrix of size (D, N), where D is the dimension of the system and N is the number of path points. The physical time of the path is specified by T, such that the time step between consecutive path points is $\Delta t = T/N$.

Geometric minimum action method (gMAM)

Minimization of the geometric action following Heymann and Vanden-Eijnden [12] and Heymann and Vanden-Eijnden [13]. gMAM reformulates MAM to avoid double optimization of both the action and the transition time. It achieves this by using a geometric action functional that is independent of the time parametrization of the path. This reparameterization invariance makes the method more robust and computationally efficient, particularly for multiscale systems.

CriticalTransitions.geometric_min_action_method — Function

geometric_min_action_method(
    sys::ContinuousTimeDynamicalSystem,
    x_i,
    x_f;
    N,
    kwargs...
) -> MinimumActionPath{_A, _B, _C, Nothing, Nothing, Nothing, Nothing, Nothing} where {_A, _B<:Real, _C}

Computes the minimizer of the geometric Freidlin-Wentzell action based on the geometric minimum action method (gMAM), using optimizers of Optimization.jl or the original formulation by Heymann and Vanden-Eijnden Heymann and Vanden-Eijnden [12]. Only the Freidlin-Wentzell action has a geometric formulation.

To set an initial path different from a straight line, see the multiple dispatch method

geometric_min_action_method(sys::CoupledSDEs, init::Matrix, arclength::Real; kwargs...).

Keyword arguments

maxiters::Int=100: maximum number of optimization iterations before the algorithm stops

abstol=1e-8: absolute tolerance of action gradient to determine convergence

reltol=1e-8: relative tolerance of action gradient to determine convergence

method = Adam(): minimization algorithm (see below)

ϵ=0.1: step size parameter in gradient descent HeymannVandenEijnden implementation

verbose=false: if true, print additional output

show_progress=true: if true, display a progress bar

Minimization algorithms

The method keyword argument takes solver methods of the Optimization.jl package; alternatively, the option method = "HeymannVandenEijnden" implements the original gMAM algorithm [12].

geometric_min_action_method(
    sys::ContinuousTimeDynamicalSystem,
    init::Matrix;
    maxiters,
    abstol,
    reltol,
    method,
    AD,
    ϵ,
    verbose,
    show_progress
) -> MinimumActionPath{_A, _B, _C, Nothing, Nothing, Nothing, Nothing, Nothing} where {_A, _B<:Real, _C}

Runs the geometric Minimum Action Method (gMAM) to find the minimum action path (instanton) from an initial condition init, given a system sys and total arc length arclength.

The initial path init must be a matrix of size (D, N), where D is the dimension of the system and N is the number of path points.

For more information see the main method, geometric_min_action_method(sys::CoupledSDEs, x_i, x_f, arclength::Real; kwargs...).

Simple geometric minimum action method (sgMAM)

Simplified minimization of the geometric action following Grafke et al. [14]. The simple gMAM reduces the complexity of the original gMAM by requiring only first-order derivatives of the underlying Hamiltonian optimization formulation. This simplifies the numerical treatment and computational complexity.

The implementation below performs a constrained gradient descent assuming an autonomous system with additive Gaussian noise.

CriticalTransitions.simple_geometric_min_action_method — Function

simple_geometric_min_action_method(
    sys::ExtendedPhaseSpace,
    x_initial::Array{T, 2};
    ϵ,
    iterations,
    show_progress,
    reltol
) -> MinimumActionPath{_A, _B, _C, Nothing, Nothing, Matrix{Float64}, Matrix{Float64}, Matrix{Float64}} where {_A, _B<:Real, _C}

Performs the simplified geometric Minimal Action Method (sgMAM) on the given system sys. Our implementation is only valid for additive noise.

This method computes the optimal path in the phase space of a Hamiltonian system that minimizes the Freidlin–Wentzell action. The Hamiltonian functions H_x and H_p define the system's dynamics in a doubled phase. The initial state x_initial is evolved iteratively using constrained gradient descent with step size parameter ϵ over a specified number of iterations. The method can display a progress meter and will stop early if the relative tolerance reltol is achieved.

The function returns a tuple containing the final state, the action value, the Lagrange multipliers, the momentum, and the state derivatives. The implementation is based on the work of Grafke et al. (2019).

CriticalTransitions.ExtendedPhaseSpace — Type

A structure representing a extanded phase space system where your dissipative vector field is embedded in a doubled dimensional phase space. Given old phase space coordinates x of a vector field f(x), we can define the canonical momenta p, such that the new phase space coordinates are (x, p). The dynamics in this extended phase space are then governed by the Hamtiltonian system:

$H = p^2 + x \dot f(x)$

Hence, this system operates in an extended phase space where the Hamiltonian is assumed to be quadratic in the extended momentum.

The struct ExtendedPhaseSpace holds the Hamilton's functions H_x and H_p.

MinimumActionPath

(gMAM) and (sgMAM) return their output as a MinimumActionPath type:

CriticalTransitions.MinimumActionPath — Type

struct MinimumActionPath{D, T<:Real, V, Phis, Ahis, Lambda, PV, GPV}

The minimum action path between two points in a D-dimensional phase space.

Fields

path::StateSpaceSet{D, T, V} where {D, T<:Real, V}: The path matrix.

action::Real: The action value associated to the path.

path_history::Any: The history of action of the paths in the optimisation algorithm (optional).

action_history::Any: The history of action of the paths in the optimisation algorithm (optional).

λ::Any: The Lagrange multiplier parameter for the minimum action path (optional).

generalized_momentum::Any: The generalized momentum of the phase space variables (optional).

path_velocity::Any: The path velocity (optional).

Constructors

MinimumActionPath(
    path,
    action;
    path_history,
    action_history,
    λ,
    generalized_momentum,
    path_velocity
)

defined at /home/runner/work/CriticalTransitions.jl/CriticalTransitions.jl/src/largedeviations/MinimumActionPath.jl:29.

Action functionals

Freidlin-Wentzell action

CriticalTransitions.fw_action — Function

fw_action(sys::CoupledSDEs, path, time) -> Any

Calculates the Freidlin-Wentzell action of a given path with time points time in a drift field specified by the deterministic dynamics f = dynamic_rule(sys) and (normalized) noise covariance matrix covariance_matrix(sys).

The path must be a (D x N) matrix, where D is the dimensionality of the system sys and N is the number of path points. The time array must have length N.

Returns a single number, which is the value of the action functional

$S_T[\phi_t] = \frac{1}{2} \int_0^T || \dot \phi_t - f(\phi_t) ||^2_Q \text{d}t$

where $\phi_t$ denotes the path in state space, $b$ the drift field, and $T$ the total time of the path. The subscript $Q$ refers to the generalized norm $||a||_Q^2 := \langle a, Q^{-1} b \rangle$ (see anorm). Here $Q$ is the noise covariance matrix normalized by $D/L_1(Q)$, with $L_1$ being the L1 matrix norm.

Geometric Freidlin-Wentzell action

CriticalTransitions.geometric_action — Function

geometric_action(sys::CoupledSDEs, path) -> Any
geometric_action(sys::CoupledSDEs, path, arclength) -> Any

Calculates the geometric action of a given path with specified arclength for the drift field specified by the deterministic dynamics f = dynamic_rule(sys) and (normalized) noise covariance matrix covariance_matrix(sys).

For a given path $\varphi$, the geometric action $\bar S$ corresponds to the minimum of the Freidlin-Wentzell action $S_T(\varphi)$ over all travel times $T>0$, where $\varphi$ denotes the path's parameterization in physical time (see fw_action). It is given by the integral

$\bar S[\varphi] = \int_0^L \left( ||\varphi'||_Q \, ||f(\varphi)||_Q - \langle \varphi', \, f(\varphi) \rangle_Q \right) \, \text{d}s$

where $s$ is the arclength coordinate, $L$ the arclength, $f$ the drift field, and the subscript $Q$ refers to the generalized dot product $\langle a, b \rangle_Q := a^{\top} \cdot Q^{-1} b$ (see anorm). Here $Q$ is the noise covariance matrix normalized by $D/L_1(Q)$, with $L_1$ being the L1 matrix norm.

Returns the value of the geometric action $\bar S$.

Onsager-Machlup action

CriticalTransitions.om_action — Function

om_action(sys::CoupledSDEs, path, time, noise_strength)

Calculates the Onsager-Machlup action of a given path with time points time for the drift field f = dynamic_rule(sys) at given noise_strength.

The path must be a (D x N) matrix, where D is the dimensionality of the system sys and N is the number of path points. The time array must have length N.

Returns a single number, which is the value of the action functional

$S^{\sigma}_T[\phi_t] = \frac{1}{2} \int_0^T \left( || \dot \phi_t - f(\phi_t) ||^2_Q + \sigma^2 \nabla \cdot f \right) \, \text{d} t$

where $\phi_t$ denotes the path in state space, $b$ the drift field, $T$ the total time of the path, and $\sigma$ the noise strength. The subscript $Q$ refers to the generalized norm $||a||_Q^2 := \langle a, Q^{-1} b \rangle$ (see anorm). Here $Q$ is the noise covariance matrix normalized by $D/L_1(Q)$, with $L_1$ being the L1 matrix norm.

For convenience, a general action function is available where the type of functional is set as an argument:

CriticalTransitions.action — Function

action(
    sys::CoupledSDEs,
    path::Matrix,
    time,
    functional;
    noise_strength
) -> Any

Computes the action functional specified by functional for a given CoupledSDEs sys and path parameterized by time.

functional = "FW": Returns the Freidlin-Wentzell action (fw_action)

functional = "OM": Returns the Onsager-Machlup action (om_action)

String method

The string method is a technique for finding transition paths between two states in a dynamical system E et al. [15]. The method represents the path as a "string" of points that connects the states and evolves it to minimize the drift along the path. The resulting tangent path is parallel to the drift of the system, i.e., the string method computes the heteroclinic orbit. For non-gradient systems (detailed -balance is broken), the heteroclinic orbit differs from the transition path, it does correctly predict, it correctly captures the deterministic dynamics from the saddle point onward ("downhill" portion of the path).

CriticalTransitions.string_method — Function

string_method(
    sys::Union{Function, ExtendedPhaseSpace},
    x_initial::Matrix;
    ϵ,
    alg,
    iterations,
    show_progress
) -> Any

Compute the string method for a given system using the method of E et al. [15].

The string method is an iterative algorithm used to find minimum energy path (MEP) between two points in phase space. It works by evolving a discretized path (string) according to the system's drift while maintaining equal arc-length parametrization between points.

This implementation allows for computation between arbitrary points, not just stable fixed points.

References

[15]

Arguments

sys::ExtendedPhaseSpace: The doubled phase space system for which the string method is computed

x_initial: Initial path discretized as a matrix where each column represents a point on the path

ϵ::Real: Step size for the evolution step

alg: SciML integrator algorithm (e.g. Euler(), Tsit5()). Defaults to Euler().

iterations::Int64: Maximum number of iterations for path convergence

show_progress::Bool: Whether to display a progress meter during computation

Returns

x: The final converged path representing the MEP

string_method(
    sys::ContinuousTimeDynamicalSystem,
    init;
    kwargs...
) -> Any

Compute the string method for a given system using the method of E et al. [15].

The string method is an iterative algorithm used to find minimum energy path (MEP) between two points in phase space. It works by evolving a discretized path (string) according to the system's drift while maintaining equal arc-length parametrization between points.

This implementation allows for computation between arbitrary points, not just stable fixed points.

References

[15]

Arguments

sys::CoupledSDEs: The system for which the string method is computed

x_initial: Initial path discretized as a matrix where each column represents a point on the path

ϵ::Real: Step size for the evolution step

alg: SciML integrator algorithm (e.g. Euler(), Tsit5()). Defaults to Euler().

iterations::Int64: Maximum number of iterations for path convergence

show_progress::Bool: Whether to display a progress meter during computation

Returns

x: The final converged path representing the MEP