Sampling transitions

... by direct simulation

These functions generate noise-induced transitions between an initial and final state.

CriticalTransitions.transition — Method

transition(
    sys::CoupledSDEs,
    x_i,
    x_f;
    rad_i,
    rad_f,
    tmax,
    cut_start,
    rad_dims,
    kwargs...
) -> Tuple{Any, Any, Any}

Generates a sample transition from point x_i to point x_f.

This function simulates sys in time, starting from initial condition x_i, until entering a length(sys.u)-dimensional ball of radius rad_f around x_f.

Keyword arguments

rad_i=0.1: radius of ball around x_i

rad_f=0.1: radius of ball around x_f

cut_start=true: if false, returns the whole trajectory up to the transition

dt=0.01: time step of integration

tmax=1e3: maximum time when the simulation stops even x_f has not been reached

rad_dims=1:length(sys.u): the directions in phase space to consider when calculating the radii rad_i and rad_f. Defaults to all directions. To consider only a subspace of state space, insert a vector of indices of the dimensions to be included.

solver=EM(): numerical solver. Defaults to Euler-Mayurama.

progress: shows a progress bar with respect to tmax

Output

[path, times, success]

path (Matrix): transition path (size [dim × N], where N is the number of time points)

times (Vector): time values (since start of simulation) of the path points (size N)

success (bool): if true, a transition occured (i.e. the ball around x_f has been reached), else false

kwargs...: keyword arguments passed to simulate

See also transitions, simulate.

CriticalTransitions.transitions — Function

transitions(
    sys::CoupledSDEs,
    x_i,
    x_f;
    ...
) -> CriticalTransitions.TransitionPathEnsemble
transitions(
    sys::CoupledSDEs,
    x_i,
    x_f,
    N;
    rad_i,
    rad_f,
    tmax,
    Nmax,
    cut_start,
    rad_dims,
    savefile,
    showprogress,
    kwargs...
) -> CriticalTransitions.TransitionPathEnsemble

Generates an ensemble of N transition samples of sys from point x_i to point x_f.

This function repeatedly calls the transition function to efficiently generate an ensemble of transitions, which are saved to a file or returned as an array of paths. Multi-threading is enabled.

Keyword arguments

rad_i=0.1: radius of ball around x_i

rad_f=0.1: radius of ball around x_f

cut_start=true: if false, returns the whole trajectory up to the transition

Nmax: number of attempts before the algorithm stops even if less than N transitions occurred.

dt=0.01: time step of integration

tmax=1e3: maximum time when the simulation stops even x_f has not been reached

rad_dims=1:length(sys.u): the directions in phase space to consider when calculating the radii rad_i and rad_f. Defaults to all directions. To consider only a subspace of state space, insert a vector of indices of the dimensions to be included.

progress: shows a progress bar with respect to Nmax

savefile: if nothing, no data is saved to a file. To save to a file, see below.

See also transition.

Saving data to file

The savefile keyword argument allows saving the data to a .jld2 or .h5 file. To do so:

Create and open a file by typing file = jld2open("filename.jld2", "a+") or file = h5open("filename.h5", "cw"). This requires JLD2.jl/HDF5.jl; the convenience functions make_jld2, make_h5 provide this out of the box.

Pass the label file to the savefile argument of transitions.

Don't forget to close(file) at the end.

Output

[samples, times, idx, N_fail]

samples (Array of Matrices): sample paths. Each path i has size (dim × Ni), where Ni is the number of path points

times (Array of Vectors): time values (since simulation start) of path points for each path

idx (Array): list of sample indices i that produced a transition

N_fail (Int): number of samples that failed to produce a transition

An example script using transitions is available here.

... in pathspace