Sampling transitions

... by direct simulation

These functions generate noise-induced transitions between an initial and final state.

CriticalTransitions.transition — Method

transition(sys::StochSystem, x_i::State, x_f::State; kwargs...)

Generates a sample transition from point x_i to point x_f.

This function simulates sys in time, starting from initial condition x_i, until entering a length(sys.u)-dimensional ball of radius rad_f around x_f.

Keyword arguments

rad_i=0.1: radius of ball around x_i

rad_f=0.1: radius of ball around x_f

cut_start=true: if false, returns the whole trajectory up to the transition

dt=0.01: time step of integration

tmax=1e3: maximum time when the simulation stops even x_f has not been reached

rad_dims=1:length(sys.u): the directions in phase space to consider when calculating the radii rad_i and rad_f. Defaults to all directions. To consider only a subspace of state space, insert a vector of indices of the dimensions to be included.

solver=EM(): numerical solver. Defaults to Euler-Mayurama.

progress: shows a progress bar with respect to tmax

Output

[path, times, success]

path (Matrix): transition path (size [dim × N], where N is the number of time points)

times (Vector): time values (since start of simulation) of the path points (size N)

success (bool): if true, a transition occured (i.e. the ball around x_f has been reached), else false

kwargs...: keyword arguments passed to simulate

See also transitions, simulate.

CriticalTransitions.transitions — Function

transitions(sys::StochSystem, x_i::State, x_f::State, N=1; kwargs...)

Generates an ensemble of N transition samples of sys from point x_i to point x_f.

This function repeatedly calls the transition function to efficiently generate an ensemble of transitions, which are saved to a file or returned as an array of paths. Multi-threading is enabled.

Keyword arguments

rad_i=0.1: radius of ball around x_i

rad_f=0.1: radius of ball around x_f

cut_start=true: if false, returns the whole trajectory up to the transition

Nmax: number of attempts before the algorithm stops even if less than N transitions occurred.

dt=0.01: time step of integration

tmax=1e3: maximum time when the simulation stops even x_f has not been reached

rad_dims=1:length(sys.u): the directions in phase space to consider when calculating the radii rad_i and rad_f. Defaults to all directions. To consider only a subspace of state space, insert a vector of indices of the dimensions to be included.

solver=EM(): numerical solver. Defaults to Euler-Mayurama

progress: shows a progress bar with respect to Nmax

savefile: if nothing, no data is saved to a file. To save to a file, see below.

See also transition.

Saving data to file

The savefile keyword argument allows saving the data to a .jld2 or .h5 file. To do so:

Create and open a file by typing file = jld2open("filename.jld2", "a+") or file = h5open("filename.h5", "cw"). This requires JLD2.jl/HDF5.jl; the convenience functions make_jld2, make_h5 provide this out of the box.

Pass the label file to the savefile argument of transitions.

Don't forget to close(file) at the end.

Output

[samples, times, idx, N_fail]

samples (Array of Matrices): sample paths. Each path i has size (dim × Ni), where Ni is the number of path points

times (Array of Vectors): time values (since simulation start) of path points for each path

idx (Array): list of sample indices i that produced a transition

N_fail (Int): number of samples that failed to produce a transition

An example script using transitions is available here.

... in pathspace

CriticalTransitions.langevinmcmc — Method

langevinmcmc(sys::StochSystem, init; kwargs...)

Runs Langevin MCMC bridge sampling for a given system sys and initial path init, using Symbolics.jl to compute the functional derivative.

Returns a three-dimensional array of size (M, length(sys.u), N), where M is the number of virtual time steps, N the number of physical time steps, and length(sys.u) the number of system dimensions.

To be further documented in following versions.

Keyword arguments

T = 15: physical path time

tmax = 250: virtual time duration

Δt = 1e-3: virtual time step

noise = sys.σ: noise strength of additive spatio-temporal noise term

annealing = false: whether to enable simulated annealing

showprogress = true: whether to print a progress bar

CriticalTransitions.stochastic_bridge — Method

stochastic_bridge(sys::StochSystem, Tphys::Float64, Δz::Float64)

Given a StochSystem sys, this function returns another higher-dimensional StochSystem that enables out-of-the-box Langevin-MCMC-type analysis. The idea is that one can transform the Langevin MCMC SPDE problem into a SDE problem of dimension $M :=(N+1)d$ that the StochSystem struct can nicely describe. Here $d$ is the dimension of sys and $N+1$ is the number of the discrete path points on your transition path (in the context of the Langevin MCMC problem).

Function arguments

sys::StochSystem: the StochSystem you are carrying out your explorations on.

Tphys::Float64: the total physical time you want to assign to your transition paths.

Δz::Float64: the physical time-step you wish to have across your transitions.

The f-field of the returned StochSystem is namely the langevinmcmc_spde function, and all other fields are chosen appropriately for the conversion between the SPDE boundary-value problem and the SDE initial-value problem. As of right now, the σ-field of the returned StochSystem equals sys.σ√(2/Δz), which is slightly different to how it is written in the original SPDE (see langevinmcmc_spde(u, p, t)); this modification is for numerical purposes, in order to account for the spatial-temporal white-noise in the original SPDE problem.

CriticalTransitions.symbolise_spde — Method

symbolise_spde(sys::StochSystem)

Given a StochSystem sys, this function uses Symbolics.jl to compute symbolic versions of multiple terms that show up in the Langevin MCMC SPDE (see langevinmcmc_spde(u, p, t)).

In a vector of length four it returns the state variables of sys, the Jacobian of the drift field $\nabla\boldsymbol{b}$, the term $\nabla\langle\boldsymbol{b}, \Sigma^{-1}\boldsymbol{b}\rangle$ which we refer to as grad_dot_term, and $\nabla(\nabla\cdot \boldsymbol{b})$ which we refer to as grad_div_term.

CriticalTransitions.langevinmcmc_spde — Method

langevinmcmc_spde(u, p, t)

For a StochSystem sys, given a discretised version of a transition path $\boldsymbol{x}(t,z) := \{\varphi(z): z \in [0, T_{\text{phys}}]\}$ that travels between two points in state-space in finite-time $T_{\text{phys}}$, this function uses symbolic differentiation to compute the value of the LangevinMCMC SPDE

$\frac{\partial \boldsymbol{x}}{\partial t} = \Sigma^{-1}\frac{\partial^2\boldsymbol{x}}{\partial z^2} - \big(\Sigma^{-1}\nabla \boldsymbol{b} - \nabla\boldsymbol{b}^\dagger\Sigma^{-1}\big)\frac{\partial \boldsymbol{x}}{\partial z} - \frac{1}{2}\nabla\langle\boldsymbol{b},\Sigma^{-1}\boldsymbol{b}\rangle\frac{\sigma^2}{2}\nabla\big(\nabla\cdot\boldsymbol{b}\big)+\sqrt{2}\sigma\boldsymbol{\eta}(z,\,t)$

across the range of physical time points z in the discretisation, say $z \in [k(\Delta z): k \in \{0, 1, 2, ..., N\}]$. In line with the boundary conditions of the Langevin MCMC SPDE problem, this partial derivative has value zero at the physical start and end times $z\in \{0,T_{\text{phys}}\}$. Here $\boldsymbol{b}, \sigma$ and $\Sigma$ are the drift field, noise strength and covariance matrix of sys, respectively.

Function arguments:

u: the transition path at virtual time t. Given that the number of discrete points in path space is $N+1$, and sys is $d$-dimensional, one should enter a concatenated vector of length $M:=(N+1)d$ of the form $[x_1(t,0), x_1(t,\Delta z), ..., x_1(t, N\Delta z), ....., x_d(t,0), x_d(t,\Delta z), ..., x_d(t, N\Delta z)]$.

p: the parameters required for the computation of Langevin MCMC SPDE. One should enter [sys.dim, sys.σ, sys.Σ, Δz, state_vars, jacobian, grad_dot_term, grad_div_term] where Δz is the physical time-step, and state_vars, jacobian, grad_dot_term, grad_div_term are precisely the outputs of the function symbolise_spde(sys::StochSystem).

t: the current fixed value of virtual time that the transition path corresponds to. One should enter a Float64.

The function returns an $M$-dimensional vector of partial-deriviates $\bigg[\frac{\partial x_1}{\partial t}(t,0), \frac{\partial x_1}{\partial t}(t,\Delta z), ..., \frac{\partial x_1}{\partial t}(t,N\Delta z), ....., \frac{\partial x_d}{\partial t}(t,0), \frac{\partial x_d}{\partial t}(t,\Delta z), ..., \frac{\partial x_d}{\partial t}(t,N\Delta z)\bigg]$.