Simulating the system

We provide two main functions to simulate a StochSystem forward in time:

[relax] for the system's deterministic evolution (in the absence of noise) based on the [ODEProblem] of DifferentialEquations.jl

[simulate] for stochastic dynamics based on the [SDEProblem] of DifferentialEquations.jl

Deterministic dynamics

CriticalTransitions.relax — Method

relax(sys::StochSystem, init::State; kwargs...)

Simulates the deterministic dynamics of StochSystem sys in time, starting at initial condition init.

This function integrates sys.f forward in time, using the ODEProblem functionality of DifferentialEquations.jl. Thus, relax is identical to simulate when setting the noise strength sys.σ = 0.

Keyword arguments

dt=0.01: time step of integration

tmax=1e3: total time of simulation

solver=Euler(): numerical solver. Defaults to explicit forward Euler

callback=nothing: callback condition

kwargs...: keyword arguments for solve(ODEProblem)

For more info, see ODEProblem. For stochastic integration, see simulate.

Warning: This function has only been tested for the Euler() solver.

Stochastic dynamics

CriticalTransitions.simulate — Method

simulate(sys::StochSystem, init::State; kwargs...)

Simulates the StochSystem sys forward in time, starting at initial condition init.

This function uses the SDEProblem functionality of DifferentialEquations.jl.

Keyword arguments

dt=0.01: time step of integration

tmax=1e3: total time of simulation

solver=EM(): numerical solver. Defaults to Euler-Mayurama

callback=nothing: callback condition

progress=true: shows a progress bar during simulation

kwargs...: keyword arguments for solve(SDEProblem)

For more info, see SDEProblem.

Warning: This function has only been tested for the EM() solver and out-of-place SDEFunctions.