Analyzing a system's stability properties

To use the following functionalities, you need to load ChoasTools.jl and Attractors.jl.

Fixed points

ChaosTools.fixedpoints — Function

fixedpoints(
    sys::CoupledSDEs,
    bmin::Vector,
    bmax::Vector
) -> Tuple{StateSpaceSet, Vector{Vector{ComplexF64}}, Vector{Bool}}

Returns fixed points, their eigenvalues and stability of the system sys within the state space volume defined by bmin and bmax.

This is a wrapper around the fixedpoints function of DynamicalSystems.jl.

Input

bmin (Vector): lower limits of the state space box to be considered, as a vector of coordinates

bmax (Vector): upper limits

alternatively box (IntervalBox) can replace bmin and bmax

Example: fixedpoints(sys, [-2,-1,0], [2,1,1]) finds the fixed points of the 3D system sys in a cube defined by the intervals [-2,2] × [-1,1] × [0,1].

Output

[fp, eigs, stable]

fp: StateSpaceSet of fixed points

eigs: vector of Jacobian eigenvalues of each fixed point

stable: vector of booleans indicating the stability of each fixed point (true=stable, false=unstable)

Additional methods

fixedpoints(sys::CoupedSDEs, box)

Edge tracking

The edge tracking algorithm is a simple numerical method to find the edge state or (possibly chaotic) saddle on the boundary between two basins of attraction. It is first introduced by Battelino et al. (1988) and further described by Skufca et al. (2006).

Attractors.edgetracking — Function

edgetracking(ds::CoupledSDEs, attractors::Dict; diffeq, kwars...)

Runs the edge tracking algorithm for the deterministic part of the CoupledSDEs ds.

Keyword arguments

diffeq=(;alg = Vern9(), reltol=1e-11): ODE solver settings

kwargs...: all keyword arguments of Attractors.edgetracking

Attractors.bisect_to_edge — Function

bisect_to_edge(ds::CoupledSDEs, attractors::Dict;
    u1, u2, bisect_thresh, diffeq, verbose, kwargs...)

Runs the bisect_to_edge function for the deterministic part of the CoupledSDEs ds.

Keyword arguments

u1: State 1 (default: first state in attractors)

u2: State 2 (default: second state in attractors)

bisect_thresh=1e-6: distance threshold

diffeq=(;alg = Vern9(), reltol=1e-11): ODE solver settings

verbose=false: Verbosity of output

ϵ_mapper=nothing: ϵ argument of Attractors.AttractorsViaProximity

kwargs...: Keyword arguments passed to Attractors.AttractorsViaProximity