Analyzing a system's stability properties

CriticalTransitions.equilib — Method

equilib(sys::StochSystem, state::State; kwargs...)

Returns the equilibrium solution of the system sys for given initial condition state.

Warning: This algorithm simply evolves the deterministic system forward in time until a steady-state condition is satisfied. Thus, the algorithm may output a false solution if it gets stuck in a quasi-equilibrium, or slowly evolving state. For more robust results, use fixedpoints.

Keyword arguments:

abstol = 1e-5: steady-state condition. Simulation ends when the rate of change (Euclidean distance in state space) of the state falls below abstol.

tmax = 1e5: maximum simulation time before the algorithm stops even if the steady-state condition is not reached.

dt = 0.01: time step of the ODE solver.

solver = Euler(): ODE solver used for evolving the state.

Missing docstring.

Missing docstring for fixedpoints(sys::StochSystem, bmin, bmax). Check Documenter's build log for details.

CriticalTransitions.fixedpoints — Method

fixedpoints(sys::StochSystem, box)

Returns fixed points, their eigenvalues and stability of the system sys within the state space volume box.

This is a wrapper around the fixedpoints function of DynamicalSystems.jl. For info on how to generate a box, see intervals_to_box.

Output

[fp, eigs, stable]

fp: Dataset of fixed points

eigs: vector of Jacobian eigenvalues of each fixed point

stable: vector of booleans indicating the stability of each fixed point (true=stable, false=unstable)

fixedpoints(sys, box)

defined at /home/runner/work/CriticalTransitions.jl/CriticalTransitions.jl/src/systemanalysis/stability.jl:48.