Analyzing a system's stability properties

Fixed points

CriticalTransitions.equilib — Method

equilib(sys::StochSystem, state::State; kwargs...)

Returns the equilibrium solution of the system sys for given initial condition state.

Warning: This algorithm simply evolves the deterministic system forward in time until a steady-state condition is satisfied. Thus, the algorithm may output a false solution if it gets stuck in a quasi-equilibrium, or slowly evolving state. For more robust results, use fixedpoints.

Keyword arguments:

abstol = 1e-5: steady-state condition. Simulation ends when the rate of change (Euclidean distance in state space) of the state falls below abstol.

tmax = 1e5: maximum simulation time before the algorithm stops even if the steady-state condition is not reached.

dt = 0.01: time step of the ODE solver.

solver = Euler(): ODE solver used for evolving the state.

Missing docstring.

Missing docstring for fixedpoints(sys::StochSystem, bmin, bmax). Check Documenter's build log for details.

Basins of attraction

CriticalTransitions.basins — Method

basins(sys::StochSystem, A, B, C, H)

Computes the basins of attraction of StochSystem sys on a plane spanned by the points A, B, C and limited by the box H.

CriticalTransitions.basinboundary — Method

basinboundary(X, Y, h; kwargs...)

Computes the basin boundary for given basins of attraction.