Predefined dynamical systems

FitzHugh-Nagumo system

A 2D system given by

\[\begin{aligned} \frac{du}{dt} &= \frac{1}{\epsilon} \left( -\alpha u^3 + \gamma u - \kappa v + I \right) \\ \frac{dv}{dt} &= -\beta v + u \ , \end{aligned}\]

where $\epsilon$ is the parameter of time scale separation between the state variables $u$ and $v$. The parameters $\alpha >0$, $\beta >1$, $\gamma>0$, and $\kappa>0$ are real constants, and $I$ denotes a driving term.

CriticalTransitions.FitzHughNagumo — Method

FitzHughNagumo(u, p, t)

Out-of-place definition of the FitzHugh-Nagumo system.

See also FitzHughNagumo!.

CriticalTransitions.FitzHughNagumo! — Method

FitzHughNagumo!(du, u, p, t)

In-place definition of the FitzHugh-Nagumo system.

See also FitzHughNagumo.

Modified Truscott-Brindley system

A two-dimensional predator-prey system given by

\[\begin{aligned} \frac{\text{d} P}{\text{d} t} &= rP\bigg(1-\frac{P}{K}\bigg)-a\frac{P^2}{h^2+P^2}Z \\ \frac{\text{d} Z}{\text{d} t} &= \xi\bigg(\frac{aP^2}{h^2+P^2}Z - mZ^2\bigg) \end{aligned}\]

The variables $P$ and $Z$ represent the concentration of phytoplankton (prey) and zooplankton (predator) species respectively. The system parameters can be interpreted as follows:

r: the phytoplankton growth rate

K: the carrying capacity of the phytoplankton

a: the maximal predation rate

h: the half-saturation constant

\xi: the time-scale separation between the evolution of phytoplankton and zooplankton

m: the zooplankton mortality rate

The model (cite...) is a modified version of the Truscott-Brindley system (cite ...), with a quadratic mortality term to enable bistability.

Thermohaline Circulation box models

CriticalTransitions.rooth_smooth — Method

rooth_smooth(u, p, t; kwargs...)

Rooth's 3-box model of the interhemispheric Thermohaline Circulation, smooth version.

The model is slightly modified by including a salinity restoration term and replacing the absolute value of the flow strength q with a smooth hyperbolic tangent approximation.

State vector

u = [T1, T2, T3, S1, S3]

Parameters

p = [F1, F3, lambd_T, lambd_S, tau1, tau2, tau3, sig1, sig2, sig3, smooth]

Keyword arguments

k = 1.5e-6: hydraulic constant

alpha = 1.5e-4: thermal expansion coefficient

beta = 8e-4: haline expansion coefficient

V = 2: volume ratio between equatorial and polar boxes