Predefined dynamical systems

FitzHugh-Nagumo system

A 2D system given by

\[\begin{aligned} \frac{du}{dt} &= \frac{1}{\epsilon} \left( -\alpha u^3 + \gamma u - \kappa v + I \right) \\ \frac{dv}{dt} &= -\beta v + u \ , \end{aligned}\]

where $\epsilon$ is the parameter of time scale separation between the state variables $u$ and $v$. The parameters $\alpha >0$, $\beta >1$, $\gamma>0$, and $\kappa>0$ are real constants, and $I$ denotes a driving term.

CriticalTransitions.FitzHughNagumo — Method

FitzHughNagumo(u, p, t)

Out-of-place definition of the FitzHugh-Nagumo system.

See also FitzHughNagumo!.

CriticalTransitions.FitzHughNagumo! — Method

FitzHughNagumo!(du, u, p, t)

In-place definition of the FitzHugh-Nagumo system.

See also FitzHughNagumo.

CriticalTransitions.fhn_ϵσ — Method

fhn_ϵσ(ϵ,σ)

A shortcut command for returning a StochSystem of the FitzHugh Nagumo system in a default setup with additive isotropic noise.

This setup fixes the parameters β = 3, α = γ = κ = 1, I = 0 and leaves the value of the time-scale parameter ϵ as a function argument. The prescribed noise process is additive and isotropic: the variables are peturbed by independently drawn identical Gaussian white noise realisations, with standard deviation σ (the other function argument).

Modified Truscott-Brindley system

A two-dimensional predator-prey system given by

\[\begin{aligned} \frac{\text{d} P}{\text{d} t} &= rP\bigg(1-\frac{P}{K}\bigg)-\frac{aP^2}{h^2+P^2}Z \\ \frac{\text{d} Z}{\text{d} t} &= \xi\bigg(\frac{aP^2}{h^2+P^2}Z - mZ^2\bigg) \end{aligned}\]

The variables $P$ and $Z$ represent the concentration of phytoplankton (prey) and zooplankton (predator) species respectively. The system parameters can be interpreted as follows:

$r$: the phytoplankton growth rate

$K$: the carrying capacity of the phytoplankton

$a$: the maximal predation rate

$h$: the half-saturation constant

$\xi$: the time-scale separation between the evolution of phytoplankton and zooplankton

$m$: the zooplankton mortality rate

The model (as discussed here) is a modified version of the original 1994 Truscott-Brindley system (see here), with a quadratic mortality term to enable bistability. Following a non-dimensionalisation of the system, the dynamical equations convert to:

\[\begin{aligned} \displaystyle\frac{\text{d} \tilde{P}}{\text{d} \tau} &= \alpha \tilde{P}\bigg(1-\beta\frac{\tilde{P}}{P_1}\bigg)-\frac{\gamma (\tilde{P}/\sqrt{P_1})^2(\tilde{Z}/Z_1)}{1+(\tilde{P}/P_1)^2} \\ \displaystyle\frac{\text{d} \tilde{Z}}{\text{d} \tau} &= \xi\tilde{Z}\bigg(\frac{(\tilde{P}/P_1)^2}{1+(\tilde{P}/P_1)^2} - \frac{\tilde{Z}}{Z_1}\bigg) \end{aligned}\]

Specifically, here we introduce the dimensionless variables $\tilde{P}, \tilde{Z}$ and $\tau$ according to the following transformations:

\[\begin{aligned} \frac{\tilde{P}}{P_1} = \frac{P}{P_0}, && \frac{\tilde{Z}}{Z_1} = \frac{Z}{Z_0}, && \tau = \frac{t}{t_0} \end{aligned}\]

where $P_1, P_0, Z_1, Z_0$ and $t_0$ are constants to be determined in the non-dimensionalisation process. We take $P_0 = h, Z_0 = a/m, t_0 = 1/a$ and subsequently introduce the following dimensionless parameters $\alpha := r/a, \beta := h/k, \gamma := a/(mh)$. The values of the (dimensionless) constants $P_1$ and $Z_1$ are tuned according to the user's preference; they can be chosen in a way such that the fixed points of the system are all contained within the $[0,1]\times [0,1]$ subspace, for instance.

Below, the functions modifiedtruscottbrindley and modifiedtruscottbrindley! implement the non-dimensional form of the system.

CriticalTransitions.modifiedtruscottbrindley — Method

modifiedtruscottbrindley(u, p, t)

Out-of-place definition of the modified Truscott-Brindley system.

See also modifiedtruscottbrindley!.

CriticalTransitions.modifiedtruscottbrindley! — Method

modifiedtruscottbrindley!(du, u, p, t)

In-place definition of the modified Truscott-Brindley system.

See also modifiedtruscottbrindley.

CriticalTransitions.modtb_αξσ — Method

modtb_αξσ(α, ξ, σ)

A shortcut command for returning a StochSystem of the modified Truscott-Brindley system in a default setup with multiplicative anisotropic noise.

This setup fixes the parameters β = 5/112, γ = 112/2.3625, P₁ = β, Z₁ = 5/6 and leaves the values of the parameters α and ξ as function arguments. The prescribed noise process is multiplicative and anisotropic: the first variable is peturbed by Gaussian white noise realisations that are multiplied by the variable's current value; the second variable has no stochastic component. The noise strength σ is left as the remaining function argument.

Thermohaline Circulation box models

CriticalTransitions.stommel — Method

stommel(u, p, t; kwargs...)

Stommel's hemispheric 2-box model of the Thermohaline Circulation (Stommel 1961).

Parameters

p = [[delta, mu, R]]

Keyword arguments

smooth = 1e6: Approximates $|x| \approx x \tanh(sx)$, where $s =$ smooth. If smooth=0, the exact absolute value is taken, making the equations non-smooth.

flow_law: Relation between density difference and flow strength. Options:

"abs": Original Stommel model [Stommel 1961]

"diffu_abs": Absolute flow law Q2 in eq. (2.4) of Cessi 1994[1]

"diffu_sqr": Absolute flow law Q3 in eq. (2.4) of Cessi 1994[1]

See also cessi.

CriticalTransitions.cessi — Method

cessi(u, p, t)

Cessi's hemispheric 2-box model of the Thermohaline Circulation (Cessi (1994)).

Parameters

p = [[alpha, mu2, pflux]]

See also stommel.

CriticalTransitions.rooth_smooth — Method

rooth_smooth(u, p, t; kwargs...)

Rooth's 3-box model of the interhemispheric Thermohaline Circulation, smooth version.

The model is slightly modified by including a salinity restoration term and replacing the absolute value of the flow strength q with a smooth hyperbolic tangent approximation.

State vector

u = [T1, T2, T3, S1, S3]

Parameters

p = [F1, F3, lambd_T, lambd_S, tau1, tau2, tau3, sig1, sig2, sig3, smooth]

Keyword arguments

k = 1.5e-6: hydraulic constant

alpha = 1.5e-4: thermal expansion coefficient

beta = 8e-4: haline expansion coefficient

V = 2: volume ratio between equatorial and polar boxes

1Cessi (1994)