Convenience functions and types

Interface to DynamicalSystems.jl

Using the functionality of DynamicalSystems.jl

A StochSystem can easily be turned into a CoupledODEs instance of DynamicalSystems.jl using the function CoupledODEs. Vice vera, a CoupledODEs system can be converted into a StochSystem via StochSystem(ds::CoupledODEs).

This way, many of the methods in DynamicalSystems.jl can be used directly, even if we have not written an analogue method that takes StochSystem as input. For example, the Lyapunov spectrum of a StochSystem, here exemplified by the FitzHugh-Nagumo model, can be computed by typing:

using CriticalTransitions, DynamicalSystems: lyapunovspectrum
sys = StochSystem(fitzhugh_nagumo, [1.,3.,1.,1.,1.,0.], zeros(2))
ls = lyapunovspectrum(CoupledODEs(sys), 10000)

DynamicalSystemsBase.CoupledODEs — Method

CoupledODEs(sys::StochSystem; diffeq, t0=0.0)

Converts a StochSystem into CoupledODEs from DynamicalSystems.jl.

CriticalTransitions.attractor_mapper — Function

attractor_mapper(sys::StochSystem, attractors, eps=0.01; kwargs...)

Wrapper around the AttractorsViaProximity mapper from DynamicalSystems.jl. Takes as input sys of type StochSystem, a vector attractors containing as elements the state vectors of the stable fixed points, and a parameter eps to control the mapping algorithm. For more info, see the docs.

StochSystem utility functions

CriticalTransitions.drift — Method

drift(sys::StochSystem, x::State)

Returns the drift field $b(x)$ of the StochSystem sys at the state vector x.

CriticalTransitions.sys_info — Method

sys_info(sys::StochSystem)

Prints StochSystem info of sys in structured format.

CriticalTransitions.sys_string — Method

sys_string(sys::StochSystem; verbose=true)

Returns StochSystem info of sys as a string.

Keyword arguments

verbose: if true, the string includes additional descriptions.

Noise functions for sys.g

CriticalTransitions.idfunc — Method

idfunc(u, p, t)

Identity function for noise function StochSystem.g (out-of-place).

CriticalTransitions.idfunc! — Method

idfunc!(du, u, p, t)

Identity function for noise function StochSystem.g (in-place).

Saving data

CriticalTransitions.make_jld2 — Function

make_jld2(text::String, relpath::String="")

Creates/opens a .jld2 file with filename of the format "ddmmyy_text.jld2". Relative file path specified by relpath (must end with /).

See make_h5 for generating a .h5 file.

CriticalTransitions.make_h5 — Function

make_h5(text::String, relpath::String="")

Creates/opens a .h5 file with filename of the format "ddmmyy_text.h5". Relative file path specified by relpath (must end with /).

See make_jld2 for generating a .jld2 file.

Miscellaneous

CriticalTransitions.is_iip — Method

is_iip(f::Function)

Asserts if f is in-place (true) or out-of-place (false).

Warning: This function simply checks if there is a ! in the function name. Thus, if you do not add a ! to in-place function names (as recommended by the Julia style guide), this test will not work.

Generalized vector norms

CriticalTransitions.anorm — Method

anorm(vec, A; square=false)

Calculates the generalized $A$-norm of the vector vec, $||v||_A := \sqrt(v^\top \cdot A \cdot v)$, where A is a square matrix of dimension (lenth(vec) x length(vec)).

Keyword arguments

square: if true, returns square of norm; else, returns norm.

CriticalTransitions.subnorm — Method

subnorm(vec; kwargs...)

Returns the Euclidean norm of the vector vec; however, if directions are specified, the norm is only computed along the specified components of the vector, i.e. in a subspace of dimension length(directions).

Keyword arguments

directions: a vector containing the indices of the components of vec to

be included. Defaults to 1:length(vec), i.e. all components.

Example

subnorm([3,7,4]; directions=[1,3] calculates the norm of only the 1st and 3rd components of the vector [3,7,4]:

$\sqrt{3^2+4^2} = 5$.

length(sys.u)-dimensional box

CriticalTransitions.intervals_to_box — Method

intervals_to_box(bmin::Vector, bmax::Vector)

Generates a box from specifying the interval limits in each dimension.

bmin (Vector): lower limit of the box in each dimension

bmax (Vector): upper limit

Example

intervals_to_box([-2,-1,0], [2,1,1]) returns a 3D box of dimensions [-2,2] × [-1,1] × [0,1].

Custom types

Parameters = Union{Vector{Float64}, Vector{Int64}, Nothing}

CovMatrix = Union{Matrix, UniformScaling{Bool}, Diagonal{Bool, Vector{Bool}}}

State = Union{Vector, SVector}