Convenience functions and types

Functions

StochSystem utility functions

CriticalTransitions.tocds — Method

tocds(sys::StochSystem; state=zeros(sys.dim))

Converts a StochSystem into a ContinuousDynamicalSystem of DynamicalSystems.jl.

Missing docstring.

Missing docstring for σg(sys::StochSystem). Check Documenter's build log for details.

Missing docstring.

Missing docstring for p(sys::StochSystem). Check Documenter's build log for details.

CriticalTransitions.sys_info — Method

sys_info(sys::StochSystem)

Prints StochSystem info of sys in structured format.

CriticalTransitions.sys_string — Method

sys_string(sys::StochSystem; verbose=true)

Returns StochSystem info of sys as a string.

Keyword arguments

verbose: if true, the string includes additional descriptions.

Noise functions for sys.g

CriticalTransitions.idfunc — Method

idfunc(u, p, t)

Identity function for noise function StochSystem.g (out-of-place).

CriticalTransitions.idfunc! — Method

idfunc!(du, u, p, t)

Identity function for noise function StochSystem.g (in-place).

sys.dim-dimensional box

CriticalTransitions.intervals_to_box — Method

intervals_to_box(bmin::Vector, bmax::Vector)

Generates a box from specifying the interval limits in each dimension.

bmin (Vector): lower limit of the box in each dimension

bmax (Vector): upper limit

Example

intervals_to_box([-2,-1,0], [2,1,1]) returns a 3D box of dimensions [-2,2] × [-1,1] × [0,1].

Saving data

CriticalTransitions.make_jld2 — Function

make_jld2(text::String, relpath::String="")

Creates/opens a .jld2 file with filename of the format "ddmmyy_text.jld2". Relative file path specified by relpath (must end with /).

See make_h5 for generating a .h5 file.

CriticalTransitions.make_h5 — Function

make_h5(text::String, relpath::String="")

Creates/opens a .h5 file with filename of the format "ddmmyy_text.h5". Relative file path specified by relpath (must end with /).

See make_jld2 for generating a .jld2 file.

Miscellaneous

CriticalTransitions.is_iip — Method

is_iip(f::Function)

Asserts if f is in-place (true) or out-of-place (false).

Warning: This function simply checks if there is a ! in the function name. Thus, if you do not add a ! to in-place function names (as recommended by the Julia style guide), this test will not work.

Types

Parameters = Union{Vector{Float64}, Vector{Int64}, Nothing}

CovMatrix = Union{Matrix, UniformScaling{Bool}, Diagonal{Bool, Vector{Bool}}}

State = Union{Vector, SVector}