Utility functions

CoupledSDEs utility functions

CriticalTransitions.covariance_matrix — Function

covariance_matrix(ds::CoupledSDEs)

Returns the covariance matrix of the stochastic term of the CoupledSDEs ds, provided that the diffusion function g can be expressed as a constant invertible matrix. If this is not the case, returns nothing.

See also diffusion_matrix.

CriticalTransitions.diffusion_matrix — Function

diffusion_matrix(ds::CoupledSDEs)

Returns the diffusion matrix of the stochastic term of the CoupledSDEs ds, provided that the diffusion function g can be expressed as a constant invertible matrix. If this is not the case, returns nothing.

Note: The diffusion matrix $Σ$ is the square root of the noise covariance matrix $Q$ (see covariance_matrix), defined via the Cholesky decomposition $Q = Σ Σ^\top$.

DynamicalSystemsBase.dynamic_rule — Function

dynamic_rule(ds::DynamicalSystem) → f

Return the dynamic rule of ds. This is never mutated and is set when initializing ds.

CriticalTransitions.noise_process — Function

noise_process(sys::CoupledSDEs) -> Any

Fetches the stochastic process $\mathcal{N}$ specified in the intergrator of sys. Returns the type DiffEqNoiseProcess.NoiseProcess.

DynamicalSystemsBase.current_state — Function

current_state(ds::DynamicalSystem) → u::AbstractArray

Return the current state of ds. This state is mutated when ds is mutated. See also initial_state, observe_state.

DynamicalSystemsBase.set_state! — Function

set_state!(ds::DynamicalSystem, u::AbstractArray{<:Real})

Set the state of ds to u, which must match dimensionality with that of ds. Also ensure that the change is notified to whatever integration protocol is used.

set_state!(ds::DynamicalSystem, value::Real, i) → u

Set the ith variable of ds to value. The index i can be an integer or a symbolic-like index for systems that reference a ModelingToolkit.jl model. For example:

i = :x # or `1` or `only(@variables(x))`
set_state!(ds, 0.5, i)

Warning: this function should not be used with derivative dynamical systems such as Poincare/stroboscopic/projected dynamical systems. Use the method below to manipulate an array and give that to set_state!.

set_state!(u::AbstractArray, value, index, ds::DynamicalSystem)

Modify the given state u and leave ds untouched.

set_state!(ds::DynamicalSystem, mapping::AbstractDict)

Convenience version of set_state! that iteratively calls set_state!(ds, val, i) for all index-value pairs (i, val) in mapping. This is useful primarily in two cases:

to partially set only some state variables,

to set variables by name (if the system is created via ModelingToolkit.jl)

so that you don't have to keep track of the order of the dynamic variables.

CriticalTransitions.drift — Function

drift(sys::CoupledSDEs{IIP}, x; t) -> Any

Returns the deterministic drift $f(x)$ of the CoupledSDEs sys at state x. For time-dependent systems, the time can be specified as a keyword argument t (by default t=0).

CriticalTransitions.div_drift — Function

div_drift(sys::CoupledSDEs, x; t) -> Any

Computes the divergence of the drift field $f(x)$ at state x. For time- dependent systems, the time can be specified as a keyword argument t (by default t=0).

Saving data

CriticalTransitions.make_jld2 — Function

make_jld2(
    text::String
) -> Union{JLD2.JLDFile{IOStream}, JLD2.JLDFile{JLD2.MmapIO}}
make_jld2(
    text::String,
    relpath::String
) -> Union{JLD2.JLDFile{IOStream}, JLD2.JLDFile{JLD2.MmapIO}}

Creates/opens a .jld2 file with filename of the format "ddmmyy_text.jld2". Relative file path specified by relpath (must end with /).

See make_h5 for generating a .h5 file.

CriticalTransitions.make_h5 — Function

make_h5(text::String) -> HDF5.File
make_h5(text::String, relpath::String) -> HDF5.File

Creates/opens a .h5 file with filename of the format "ddmmyy_text.h5". Relative file path specified by relpath (must end with /).

See make_jld2 for generating a .jld2 file.

length(sys.u)-dimensional box

CriticalTransitions.intervals_to_box — Method

intervals_to_box(bmin::Vector, bmax::Vector) -> Any

Generates a box from specifying the interval limits in each dimension.

bmin (Vector): lower limit of the box in each dimension

bmax (Vector): upper limit

Example

intervals_to_box([-2,-1,0], [2,1,1]) returns a 3D box of dimensions [-2,2] × [-1,1] × [0,1].